Rename/Move interface.ts -> settings/SettingsInterface.ts
